% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{B_1d}
\alias{B_1d}
\title{multivariate skewness test}
\usage{
B_1d(x)
}
\arguments{
\item{x}{A contingency table matrix.}
}
\value{
skewness coeff of x
}
\description{
Computes the multivariate skewness coeff.
}
\examples{
\dontrun{
d=2
sigma=matrix(rep(0,d*d),nrow = d)
for(i in 1:d){
  for (j in 1:d){
      sigma[i,j]=0.5^(abs(i-j))
  }
}
mu=matrix(rep(0,d),nrow = 1)
x <- mvrnorm(n[i],mu,sigma) 
B_1d(x)
}
}
