% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{count5.F}
\alias{count5.F}
\title{Count Five test and F test}
\usage{
count5.F(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{Another vector with the same length of x.}
}
\value{
The p value of the F statistic and Count Five test statistic.
}
\description{
Both calculate the F statistic and Count Five test statistic.
}
\examples{
\dontrun{
x <- rnorm(n=50, 0, 1) 
y <- rnorm(n=50, 0, 2) 
count5.F(x,y)
}
}
