% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{count5test}
\alias{count5test}
\title{Count Five test}
\usage{
count5test(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{Another vector with the same length of x.}
}
\value{
The value of a test statistic.
}
\description{
Count Five test is used to test whether the two sets of data have the same variance
}
\examples{
\dontrun{
x <- rnorm(n=50, 0, 1) 
y <- rnorm(n=50, 0, 2) 
count5test(x,y)
}
}
