% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{eqdist.nn}
\alias{eqdist.nn}
\title{evaluating the performance of the NN}
\usage{
eqdist.nn(z, sizes, k, R)
}
\arguments{
\item{z}{Sample data to be tested}

\item{sizes}{Total sample size}

\item{k}{Nearest neighbor quantity selection}

\item{R}{The number of the bootstrap}
}
\value{
KNN statistic and its P value
}
\description{
evaluating the performance of the NN used to evaluate whether the variance is homophase
}
\examples{
\dontrun{
m <- 1e3; k<-3; p<-2; set.seed(12345) 
n1 <- n2 <- 50; R<-999; n <- n1+n2; N = c(n1,n2)
x <- matrix(rnorm(n1*p),ncol=p); y <- cbind(rnorm(n2),rnorm(n2,0,1.8)); z <- rbind(x,y) 
eqdist.nn(z,N,k,R)
}
}
