% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolisC}
\alias{rwMetropolisC}
\title{Implement a random walk Metropolis sampler.}
\usage{
rwMetropolisC(mu, lambda, sigma, x0, N)
}
\arguments{
\item{mu}{The mean of Laplace distribution.}

\item{lambda}{The lambda parameter of Laplace distribution.}

\item{sigma}{The standard deviation of the norm target distribution.}

\item{x0}{The initial value of a random walk sequence.}

\item{N}{The length of a random walk sequence.}
}
\value{
a random walk sequence and the number of times it was accepted.
}
\description{
Implement a random walk Metropolis sampler for generating the standard Laplace distribution.
}
\examples{
\dontrun{
set.seed(1234)
mu=0;lambda=1;sigma=1
x0=10
N=15000
rw<- rwMetropolisC(mu,lambda, sigma, x0, N)
}
}
