% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{sk}
\alias{sk}
\title{skewness coeff}
\usage{
sk(x)
}
\arguments{
\item{x}{A contingency table matrix.}
}
\value{
skewness coeff of x
}
\description{
Computes the sample skewness coeff.
}
\examples{
\dontrun{
x=runif(30)
sk(x)
}
}
